fix(telemetry): emit int-typed registry attributes as i64
The tracing→OTel bridge visitors implement record_i64 but not
record_u64, so u64/usize span and event fields fall through to
record_debug and export as strings — violating the registry's
`type: int` declarations under weaver live-check (surfaced on #1007:
signaldb.job.{tables_processed,partitions_dropped,snapshots_expired,
bytes_reclaimed,duration_ms} arrived as strings once the monolith
started running retention in the check window).

Cast every int-typed registry attribute emission to i64 at the call
site (retention enforcer events, WAL and writer span fields), and pin
the rule in registry_pins: any `type: int` attribute assigned without
an `as i64` cast fails the drift gate.

+++ src/common/src/wal/mod.rs
@@ -778,7 +778,7 @@ impl Wal {
     #[tracing::instrument(
         level = "debug",
         skip_all,
-        fields(signaldb.wal.operation = ?operation, signaldb.wal.data_size = data.len())
+        fields(signaldb.wal.operation = ?operation, signaldb.wal.data_size = data.len() as i64)
     )]
     pub async fn append(
         &self,
@@ -1013,7 +1013,7 @@ impl Wal {
     #[tracing::instrument(
         level = "debug",
         skip_all,
-        fields(signaldb.wal.entry_count = entry_ids.len())
+        fields(signaldb.wal.entry_count = entry_ids.len() as i64)
     )]
     pub async fn mark_processed_many(&self, entry_ids: &[Uuid]) -> Result<()> {
         if entry_ids.is_empty() {
